+++ templates/wordpress/assets/js/app.js
@@ -3,10 +3,19 @@
 
 $(document).ready(start);
 $(document).on('page:load', start);
+$(window).load(startOnLoad);
 
 function start() {
   app.init();
   woo.init();
+
+  animateOnScroll.init();
+  responsiveTable.init();
+}
+
+// functions that needs to run only after everything loads
+function startOnLoad() {
+
 }
 
 /////
@@ -17,12 +26,13 @@ var app = {
   init: function() {
     this.mobileMenu();
     this.searchMenu();
-    this.pagination();
+
+    this.commentFormToggle();
   },
 
   // Toggle Mobile menu
   mobileMenu: function() {
-    $('#menu-toggle').on('click', open);
+    $('.menu-toggle').on('click', open);
     $('.menu-wrapper').on('click', preventClose);
     $(document).on('click', close);
 
@@ -47,18 +57,128 @@ var app = {
     }
   },
 
-  // Handle pagination dropdown
-  pagination: function() {
-    if($('#pagination').length <= 0) { return false; }
+  commentFormToggle: function() {
+    var replyTitle = document.getElementById('reply-title').childNodes;
+    var replyTo = replyTitle[1].childNodes[0].nodeValue;
+    var placeholder = replyTitle[0].nodeValue + (replyTo ? replyTo : '') + '…';
 
-    $('#pagination').on('change', onChange);
+    $('.comment-form textarea')
+      .attr('placeholder', placeholder)
+      .on('focus', activateForm);
 
-    function onChange(e) {
-      var url = $(this).val();
-      window.location = url;
+    function activateForm(e) {
+      $(this).closest('.comment-form').addClass('active');
     }
   }
 };
+
+/*
+  RESPONSIVE TABLE
+  - Freeze the first column on mobile.
+  - Demo: http://codepen.io/hrsetyono/pen/KaKzzZ
+*/
+
+var responsiveTable = {
+  init: function() {
+    $(".post-content table").each(this.setup);
+  },
+
+  setup: function() {
+    var $t = $(this);
+
+    // no need to responsify if only three columns
+    if($t.find('tr:first-child td, tr:first-child th').length <= 3) {
+      return false;
+    }
+
+    // wrap large table and clone it to small one
+    $t.wrap('<div class="table-large"></div>');
+    var $tSmall = $t.clone().insertAfter($t.closest('.table-large') ).wrap('<div class="table-small"></div>');
+
+    // mark the neighboring rowspan
+    var $spannedCols = $tSmall.find('[rowspan]');
+    $spannedCols.each(markSpannedColumn);
+
+    // remove width attribute that can cause formatting issue
+    $tSmall.find('[width]').removeAttr('width');
+
+    // create pinned table
+    var $tSmall2 = $tSmall.clone();
+    $tSmall2.insertAfter($tSmall);
+
+    // wrap both the small tables
+    $tSmall.wrap('<div class="table-scrollable"></div>');
+    $tSmall2.wrap('<div class="table-pinned"></div>');
+
+    /////
+
+    function markSpannedColumn() {
+      var $cell = $(this);
+      var num = $cell.attr('rowspan') - 1;
+
+      $cell.closest('tr').nextAll('tr:lt(' + num + ')')
+        .find('th:first-child, td:first-child').addClass('spanned');
+    }
+  },
+};
+
+/*
+  ANIMATE ON SCROLL
+  - Animate the element when visible, uses [data-animate] attribute
+
+  <div data-animate="fadeInUp">...</div>
+*/
+var animateOnScroll = {
+  init: function() {
+    var _this = this;
+    if($('[data-animate]').length <= 0) { return false; }
+
+    // Animate elements that already visible
+    $("[data-animate]").each(_this.run);
+
+    // Animate elements upon scrolling
+    $(window).scroll(function() {
+      $("[data-animate]").each(_this.run);
+    });
+  },
+
+  /*
+    Check whether the element is visible on screen. If yes, start animating.
+
+    @param $element DOM
+  */
+  run: function($element) {
+    var $element = $(this);
+
+    var topOfWindow = $(window).scrollTop();
+    var threshold = $(window).height() - 50;
+    var imagePos = $element.offset().top;
+
+    if (imagePos < topOfWindow + threshold) {
+      animate.startAnimating($element);
+    }
+  },
+
+  /*
+    Start the animation with delay, if any
+
+    @param $element DOM
+  */
+  startAnimating: function($element) {
+    var animationData = $element.data("animate");
+    var animation = {
+      name: animationData.match(/^\S+/),
+      delay: animationData.match(/\d+/) || 0,
+    };
+
+    var addAnimateClass = function() {
+      $element.addClass("animated " + animation.name);
+    };
+
+    setTimeout(addAnimateClass, animation.delay);
+  },
+}
+
 
 // ----- WOOCOMMERCE -----